backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
transformers==4.41.2
torch==2.3.1
requests==2.32.3
beautifulsoup4==4.12.3
joblib==1.4.2
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.7
sentence-transformers==3.1.1
chromadb==0.5.5
